# api/openapi.yaml
info:
  title: PaySuper payment solution service.
  version: "1.0.0"
  description: Swagger Specification for PaySuper Management API.
servers:
  - url: "https://api.pay.super.com"
    description: Production API
components:
  schemas:
...
    KeyProductInfo:
      type: object
      properties:
        id:
          type: string
        project_id:
          type: string
        name:
          type: string
        description:
          type: string
        long_description:
          type: string
        images:
          type: array
          items:
            type: string
        platforms:
          type: array
          items:
            properties:
              id:
                type: string
              name:
                type: string
              price:
                type: object
                properties:
                  amount:
                    type: number
                  currency:
                    type: string
                  region:
                    type: string
                  is_fallback:
                    type: boolean
...
    ResponseErrorMessage:
      type: object
      properties:
        code:
          description: "The error code."
          type: string
        message:
          description: "The error short description."
          type: string
        details:
          description: "The error details."
          type: string
...
paths:
...
  /auth/api/v1/key-products/{key_product_id}:
    get:
      summary: "Get the list of platforms and prices for the key-activated product"
      operationId: keyProductsIdPathGetKeyProduct
      description: "Get the list of platforms and prices for the key-activated product. This list can be filtered by country, language or currency."
      tags:
      - Product
      parameters:
      - name: key_product_id
        description: "The unique identifier for the key-activated product."
        in: path
        required: true
        schema:
          type: string
      - name: country
        description: "The country's name to calculate the price for."
        in: query
        required: false
        schema:
          type: string
      - name: language
        description: The language.
        in: query
        required: false
        schema:
          type: string
      - name: currency
        description: The price currency.
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: "Returns the product data (platforms, prices and others)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeyProductInfo"
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
...
openapi: "3.0.2"